refactored text component, made modal wider
