+++ client/js/lib_components/Text.js
@@ -1,12 +1,23 @@
 var React = require('react');
 var marked = require('marked');
 
+var renderer = new marked.Renderer();
+
+renderer.heading = function (text, level) {
+  var escapedText =  text.toLowerCase().split(/[^\w]+/g).slice(3).join('_');
+
+  return '<h' + level + '><a name="' +
+                escapedText +
+                 '" class="anchor" href="#' +
+                 escapedText +
+                 '"><span class="header-link"></span></a>' +
+                  text + '</h' + level + '>';
+}
+
 var Text = React.createClass({
 
     render: function(){
-        console.log(this.props.body)
-        return <div id={this.props.id}  dangerouslySetInnerHTML={ { __html: marked(this.props.body) } }/>
-        // return (<div>FOOO</div>)
+        return <div id={this.props.id} className={this.props.className} dangerouslySetInnerHTML={ { __html: marked(this.props.body) } }/>
     }
 });
